server/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cloudreactor-task-manager"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "six>=1.15.0,<2.0.0",
    "beautifulsoup4>=4.12.3,<5.0.0",
    "boto3>=1.37.18,<2.0.0",
    "cloudreactor-procwrapper[allextras]==6.0.2",
    "crontab>=1.0.1,<2.0.0",
    "cryptography==43.0.3",
    "Django>=4.2.20,<5.0.0",
    "django-anymail[amazon_ses]>=10.3,<11.0",
    "django-cors-headers>=4.7.0,<5.0.0",
    "django-csp>=3.8,<4.0.0",
    "django-dynamic-fixtures>=0.2.1,<1.0.0",
    "django-environ>=0.12.0,<1.0.0",
    "django-extensions>=3.2.3,<4.0.0",
    "django-filter>=24.2,<25.0",
    "django-guardian>=2.4.0,<3.0.0",
    "django-middleware-global-request>=0.3.5,<1.0.0",
    "django-model-utils>=4.5.0,<5.0.0",
    "django-templated-email>=3.0.1,<4.0.0",
    "djangorestframework>=3.15.2,<4.0.0",
    "djangorestframework-simplejwt>=5.3.1,<6.0.0",
    "djoser>=2.3.1,<3.0.0",
    "django-typed-models>=0.14.0",
    "django-rest-typed-models>=0.1.0",
    "drf-flex-fields>=1.0.2,<2.0.0",
    "drf-spectacular>=0.28.0,<1.0.0",
    "Jinja2>=3.1.5,<4.0.0",
    "gunicorn>=23.0.0",
    "networkx>=2.8.8,<3.0.0",
    "psycopg[binary,pool]==3.2.6,<4.0.0",
    "pdpyras>=5.2.0,<6.0.0",
    "pydantic>=1.10.21,<2.0.0",
    "pypd>=1.1.0,<2.0.0",
    "python-dateutil>=2.8.2,<3.0.0",
    "python-dotenv>=1.0.1,<2.0.0",
    "python-intervals>=1.10.0,<2.0.0",
    "pytz>=2025.1",
    "PyYAML>=6.0.2,<7.0.0",
    "requests>=2.32.3,<3.0.0",
    "sqlparse>=0.5.3,<1.0.0",
    "stringcase==1.2.0,<2.0.0",
    "whitenoise>=6.9.0"
    # "django-queryinspect==1.1.0" # Uncomment if needed
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.1",
    "pytest-assert-utils>=0.3.1",
    "pytest-cov>=5.0.0",
    "pytest-django>=4.8.0",
    "pytest-env>=1.1.3",
    "pytest-factoryboy>=2.7.0,<3.0.0",
    "factory-boy>=3.3.0,<4.0.0",
    "model_bakery>=1.18.0,<2.0.0",
    "moto[all]>=5.0.4,<6.0.0",
    "pylint==3.1.0",
    "pylint-django==2.5.5",
    "mypy>=1.7.0",
    "safety==1.10.3",
    "mypy-extensions==1.0.0",
    "django-stubs==4.2.7",
    "djangorestframework-stubs==3.14.5",
    "boto3-stubs==1.34.34",
    "mypy-boto3==1.34.34",
    "mypy-boto3-cloudformation==1.34.32",
    "mypy-boto3-lambda===1.34.0",
    "types-python-dateutil>=2.8.19.20240106"
]

[tool.uv]
# requirements = ["requirements.txt"]

[tool.setuptools.packages.find]
include = ["processes*", "task_manager*", "spectacular*", "tests*"]
exclude = ["sql*", "schema*", "static*"]

[pytest]
addopts = "--reuse-db --capture=tee-sys"
env = [
    "DJANGO_SETTINGS_MODULE=task_manager.settings",
    "IN_PYTEST=TRUE"
]
